GUI Improve Stage1 Complete
