+++ NEWorld.vcxproj
@@ -179,17 +179,13 @@
     <ClCompile Include="source\Frustum.cpp" />
     <ClCompile Include="source\GUI.cpp" />
     <ClCompile Include="source\Hitbox.cpp" />
-    <ClCompile Include="source\MainMenu.cpp" />
     <ClCompile Include="source\Menus.cpp" />
-    <ClCompile Include="source\MultiplayerGameMenu.cpp" />
     <ClCompile Include="source\Network.cpp" />
     <ClCompile Include="source\NEWorld.cpp" />
     <ClCompile Include="source\OnlinePlayer.cpp" />
-    <ClCompile Include="source\OptionsMenu.cpp" />
     <ClCompile Include="source\Particles.cpp" />
     <ClCompile Include="source\Player.cpp" />
     <ClCompile Include="source\Renderer.cpp" />
-    <ClCompile Include="source\RenderOptions.cpp" />
     <ClCompile Include="source\TextRenderer.cpp" />
     <ClCompile Include="source\Textures.cpp" />
     <ClCompile Include="source\World.cpp" />
